+++ packages/flutter/example/integration_test/native_jni_utils_test.dart
@@ -105,7 +105,7 @@ void main() {
 
     test('dartToJObject converts list (drops nulls)', () {
       using((arena) {
-        final javaList = dartToJObject(inputList).as(JList.type(JObject.type))
+        final javaList = dartToJObject(inputList).as(JList.type)
           ..releasedBy(arena);
         _expectJniList(javaList, expectedList, arena);
       });
@@ -113,8 +113,7 @@ void main() {
 
     test('dartToJObject converts map (drops null values)', () {
       using((arena) {
-        final javaMap = dartToJObject(inputMap)
-            .as(JMap.type(JString.type, JObject.type))
+        final javaMap = dartToJObject(inputMap).as(JMap.type)
           ..releasedBy(arena);
         _expectJniMap(javaMap, expectedMap, arena);
       });
@@ -129,7 +128,7 @@ void main() {
 
     test('dartToJMap', () {
       using((arena) {
-        final javaMap = dartToJMap(inputMap)..releasedBy(arena);
+        final javaMap = dartToJMap(inputMap).as(JMap.type)..releasedBy(arena);
         _expectJniMap(javaMap, expectedMap, arena);
       });
     });
@@ -162,46 +161,46 @@ void _expectJniStringEquals(JObject? javaObject, String expected) {
 void _expectJniLongEquals(JObject? javaObject, int expected) {
   expect(javaObject, isNotNull);
   final javaLong = javaObject!.as(JLong.type);
-  expect(javaLong.longValue(releaseOriginal: true), expected);
+  expect(javaLong.toDartInt(releaseOriginal: true), expected);
 }
 
 void _expectJniDoubleEquals(JObject? javaObject, double expected) {
   expect(javaObject, isNotNull);
   final javaDouble = javaObject!.as(JDouble.type);
-  expect(javaDouble.doubleValue(releaseOriginal: true), expected);
+  expect(javaDouble.toDartDouble(releaseOriginal: true), expected);
 }
 
 void _expectJniBoolEquals(JObject? javaObject, bool expected) {
   expect(javaObject, isNotNull);
   final javaBoolean = javaObject!.as(JBoolean.type);
-  expect(javaBoolean.booleanValue(releaseOriginal: true), expected);
+  expect(javaBoolean.toDartBool(releaseOriginal: true), expected);
 }
 
-JObject? _get(JMap<JString, JObject> javaMap, String key, Arena arena) =>
-    javaMap[key.toJString()..releasedBy(arena)];
+JObject? _get(JMap<JObject?, JObject?> javaMap, String key, Arena arena) =>
+    javaMap.asDart()[key.toJString()..releasedBy(arena)];
 
 void _expectJniList(
-  JList<JObject> javaList,
+  JList<JObject?> javaList,
   List<Object?> expectedListValues,
   Arena arena,
 ) {
-  expect(javaList.length, expectedListValues.length);
+  final list = javaList.asDart();
+  expect(list.length, expectedListValues.length);
 
-  _expectJniStringEquals(javaList[0], expectedListValues[0] as String);
-  _expectJniLongEquals(javaList[1], expectedListValues[1] as int);
-  _expectJniDoubleEquals(javaList[2], expectedListValues[2] as double);
-  _expectJniBoolEquals(javaList[3], expectedListValues[3] as bool);
-  _expectJniStringEquals(javaList[4], expectedListValues[4] as String);
+  _expectJniStringEquals(list[0], expectedListValues[0] as String);
+  _expectJniLongEquals(list[1], expectedListValues[1] as int);
+  _expectJniDoubleEquals(list[2], expectedListValues[2] as double);
+  _expectJniBoolEquals(list[3], expectedListValues[3] as bool);
+  _expectJniStringEquals(list[4], expectedListValues[4] as String);
 
-  final nestedList = javaList[5].as(JList.type(JObject.type))
-    ..releasedBy(arena);
+  final nestedList = list[5]!.as(JList.type)..releasedBy(arena);
+  final nestedListDart = nestedList.asDart();
   final expectedNestedList = expectedListValues[5] as List<Object?>;
-  expect(nestedList.length, expectedNestedList.length);
-  _expectJniStringEquals(nestedList[0], expectedNestedList[0] as String);
-  _expectJniLongEquals(nestedList[1], expectedNestedList[1] as int);
+  expect(nestedListDart.length, expectedNestedList.length);
+  _expectJniStringEquals(nestedListDart[0], expectedNestedList[0] as String);
+  _expectJniLongEquals(nestedListDart[1], expectedNestedList[1] as int);
 
-  final nestedMap = javaList[6].as(JMap.type(JString.type, JObject.type))
-    ..releasedBy(arena);
+  final nestedMap = list[6]!.as(JMap.type)..releasedBy(arena);
   _expectJniNestedMap(
     nestedMap,
     expectedListValues[6] as Map<String, Object?>,
@@ -211,11 +210,11 @@ void _expectJniList(
 }
 
 void _expectJniMap(
-  JMap<JString, JObject> javaMap,
+  JMap<JObject?, JObject?> javaMap,
   Map<String, Object?> expectedMapValues,
   Arena arena,
 ) {
-  expect(javaMap.length, expectedMapValues.length);
+  expect(javaMap.asDart().length, expectedMapValues.length);
 
   final expectedList = expectedMapValues['list']! as List<Object?>;
   final expectedNestedList = expectedList[5] as List<Object?>;
@@ -233,12 +232,11 @@ void _expectJniMap(
   _expectJniStringEquals(
       _get(javaMap, 'key5', arena), expectedMapValues['key5'] as String);
 
-  final nestedList = _get(javaMap, 'list', arena)!.as(JList.type(JObject.type))
+  final nestedList = _get(javaMap, 'list', arena)!.as(JList.type)
     ..releasedBy(arena);
   _expectJniList(nestedList, expectedList, arena);
 
-  final nestedMap = _get(javaMap, 'nestedMap', arena)!
-      .as(JMap.type(JString.type, JObject.type))
+  final nestedMap = _get(javaMap, 'nestedMap', arena)!.as(JMap.type)
     ..releasedBy(arena);
   _expectJniNestedMap(
       nestedMap, expectedNestedMap, expectedNestedList.length, arena);
@@ -247,7 +245,7 @@ void _expectJniMap(
 }
 
 void _expectJniNestedMap(
-  JMap<JString, JObject> javaNestedMap,
+  JMap<JObject?, JObject?> javaNestedMap,
   Map<String, Object?> expectedNestedMapValues,
   int expectedNestedListLength,
   Arena arena,
@@ -255,13 +253,13 @@ void _expectJniNestedMap(
   _expectJniStringEquals(_get(javaNestedMap, 'innerString', arena),
       expectedNestedMapValues['innerString'] as String);
 
-  final innerList = _get(javaNestedMap, 'innerList', arena)!
-      .as(JList.type(JObject.type))
+  final innerList = _get(javaNestedMap, 'innerList', arena)!.as(JList.type)
     ..releasedBy(arena);
-  expect(innerList.length, expectedNestedListLength);
-  _expectJniLongEquals(innerList[0],
+  final innerListDart = innerList.asDart();
+  expect(innerListDart.length, expectedNestedListLength);
+  _expectJniLongEquals(innerListDart[0],
       (expectedNestedMapValues['innerList']! as List<Object?>)[0] as int);
-  _expectJniLongEquals(innerList[1],
+  _expectJniLongEquals(innerListDart[1],
       (expectedNestedMapValues['innerList']! as List<Object?>)[1] as int);
 
   expect(_get(javaNestedMap, 'innerNull', arena), isNull);